requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.3
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==0.6.5
langsmith==0.4.14
matplotlib==3.10.1
mysql-connector-python==9.3.0
numpy==2.2.4
orjson==3.11.2
packaging==24.2
pandas==2.3.0
pillow==11.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.5
requests-toolbelt==1.0.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zstandard==0.24.0
